# hexabid/auctions-api-contract/target/generated-sources/openapi/src/main/java/com/acme/auctions/contract/api/AuctionsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.14.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.acme.auctions.contract.api;

import com.acme.auctions.contract.model.AuctionListResponse;
import com.acme.auctions.contract.model.AuctionResponse;
import com.acme.auctions.contract.model.AuctionSort;
import com.acme.auctions.contract.model.AuctionStatus;
import com.acme.auctions.contract.model.CreateAuctionRequest;
import com.acme.auctions.contract.model.CurrentUserProfileResponse;
import java.util.UUID;
import org.springframework.lang.Nullable;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-07T00:19:55.094556016+01:00[Europe/Warsaw]", comments = "Generator version: 7.14.0")
@Validated
public interface AuctionsApi {

    default AuctionsApiDelegate getDelegate() {
        return new AuctionsApiDelegate() {};
    }

    /**
     * GET /api/auctions : Browse auctions for search and listing screens
     *
     * @param query  (optional)
     * @param status  (optional)
     * @param sort  (optional)
     * @param limit  (optional, default to 20)
     * @param after  (optional)
     * @return Auction list page (status code 200)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/auctions",
        produces = { "application/json" }
    )
    
    default ResponseEntity<AuctionListResponse> browseAuctions(
         @Valid @RequestParam(value = "query", required = false) @Nullable String query,
         @Valid @RequestParam(value = "status", required = false) @Nullable AuctionStatus status,
         @Valid @RequestParam(value = "sort", required = false) @Nullable AuctionSort sort,
        @Min(1) @Max(50)  @Valid @RequestParam(value = "limit", required = false, defaultValue = "20") Integer limit,
         @Valid @RequestParam(value = "after", required = false) @Nullable String after
    ) {
        return getDelegate().browseAuctions(query, status, sort, limit, after);
    }


    /**
     * GET /api/me/auctions : Browse auctions created by current user
     *
     * @param status  (optional)
     * @param sort  (optional)
     * @param limit  (optional, default to 20)
     * @param after  (optional)
     * @return Current user auctions (status code 200)
     *         or Authentication required (status code 401)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/me/auctions",
        produces = { "application/json" }
    )
    
    default ResponseEntity<AuctionListResponse> browseMyAuctions(
         @Valid @RequestParam(value = "status", required = false) @Nullable AuctionStatus status,
         @Valid @RequestParam(value = "sort", required = false) @Nullable AuctionSort sort,
        @Min(1) @Max(50)  @Valid @RequestParam(value = "limit", required = false, defaultValue = "20") Integer limit,
         @Valid @RequestParam(value = "after", required = false) @Nullable String after
    ) {
        return getDelegate().browseMyAuctions(status, sort, limit, after);
    }


    /**
     * GET /api/me/bids : Browse auctions where current user placed bids
     *
     * @param status  (optional)
     * @param sort  (optional)
     * @param limit  (optional, default to 20)
     * @param after  (optional)
     * @return Current user bid auctions (status code 200)
     *         or Authentication required (status code 401)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/me/bids",
        produces = { "application/json" }
    )
    
    default ResponseEntity<AuctionListResponse> browseMyBids(
         @Valid @RequestParam(value = "status", required = false) @Nullable AuctionStatus status,
         @Valid @RequestParam(value = "sort", required = false) @Nullable AuctionSort sort,
        @Min(1) @Max(50)  @Valid @RequestParam(value = "limit", required = false, defaultValue = "20") Integer limit,
         @Valid @RequestParam(value = "after", required = false) @Nullable String after
    ) {
        return getDelegate().browseMyBids(status, sort, limit, after);
    }


    /**
     * POST /api/auctions : Create a new auction
     *
     * @param createAuctionRequest  (required)
     * @return Auction created (status code 201)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/auctions",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<AuctionResponse> createAuction(
         @Valid @RequestBody CreateAuctionRequest createAuctionRequest
    ) {
        return getDelegate().createAuction(createAuctionRequest);
    }


    /**
     * GET /api/auctions/{auctionId} : Get auction details
     *
     * @param auctionId  (required)
     * @return Auction details (status code 200)
     *         or Auction not found (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/auctions/{auctionId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<AuctionResponse> getAuctionById(
         @PathVariable("auctionId") UUID auctionId
    ) {
        return getDelegate().getAuctionById(auctionId);
    }


    /**
     * GET /api/me : Get current authenticated user profile
     *
     * @return Current user profile (status code 200)
     *         or Authentication required (status code 401)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/me",
        produces = { "application/json" }
    )
    
    default ResponseEntity<CurrentUserProfileResponse> getCurrentUserProfile(
        
    ) {
        return getDelegate().getCurrentUserProfile();
    }

}